+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Default variables can now be defined per program in the global sugarkube-conf.yaml file. Keys map to programs in a kapp's 'requires' block
 * The kubeconfig file donwloaded to access a kops cluster is now terminated once SSH port forwarding is terminated
 * Any kapps that declare outputs now need to implement an extra make target called 'output' to non-destructively write their outputs to a file. 
+* Rename the 'parallisation' option in manifests to a boolean 'is_sequential' indicating that each kapp in the manifest depends on the previous one
 
 ## 0.6.0 (25/3/19)
 * Major code clean up & refactoring